refactor: move hook to hooks folder
