+++ admin/src/lexical/plugins/CodeActionMenuPlugin/index.tsx
@@ -13,13 +13,13 @@ import './index.css';
 import { $isCodeNode, CodeNode, getLanguageFriendlyName, normalizeCodeLang } from '@lexical/code';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { $getNearestNodeFromDOMNode, isHTMLElement } from 'lexical';
-import { useEffect, useRef, useState } from 'react';
 import * as React from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 
+import { useDebounce } from '../../hooks/useDebounce';
 import { CopyButton } from './components/CopyButton';
 import { canBePrettier, PrettierButton } from './components/PrettierButton';
-import { useDebounce } from './utils';
 
 const CODE_PADDING = 8;
