+++ admin/src/lexical/plugins/TableHoverActionsPlugin/index.tsx
@@ -25,11 +25,11 @@ import {
 } from '@lexical/table';
 import { $findMatchingParent, mergeRegister } from '@lexical/utils';
 import { $getNearestNodeFromDOMNode, isHTMLElement, NodeKey } from 'lexical';
-import { useEffect, useMemo, useRef, useState } from 'react';
 import * as React from 'react';
+import { useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 
-import { useDebounce } from '../CodeActionMenuPlugin/utils';
+import { useDebounce } from '../../hooks/useDebounce';
 
 const BUTTON_WIDTH_PX = 20;
